# api/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/friends/blockSubscribeByEmail": {
            "post": {
                "description": "Requirement 5: As a user, I need an API to block updates from an email address.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend API"
                ],
                "summary": "Block subscribe by email",
                "parameters": [
                    {
                        "description": "Block updates from an email address",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BlockSubscribeRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/friends/createConnection": {
            "post": {
                "description": "Requirement 1: As a user, I need an API to create a friend connection between two email addresses.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend API"
                ],
                "summary": "Create a friend connection",
                "parameters": [
                    {
                        "description": "Create a friend connection between 2 user email",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FriendConnectionRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/friends/showCommonFriendList": {
            "post": {
                "description": "Requirement 3: As a user, I need an API to retrieve the common friends list between two email addresses.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend API"
                ],
                "summary": "Show common Friend list",
                "parameters": [
                    {
                        "description": "Retrieve the common friends list between two email addresses",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CommonFriendListRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/friends/showFriendsByEmail": {
            "post": {
                "description": "Requirement 2: As a user, I need an API to retrieve the friends list for an email address.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend API"
                ],
                "summary": "Get Friend list by email",
                "parameters": [
                    {
                        "description": "Get a list of friend by user email",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FriendListRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/friends/showSubscribingEmailListByEmail": {
            "post": {
                "description": "Requirement 6: As a user, I need an API to retrieve all email addresses that can receive updates from an email address.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend API"
                ],
                "summary": "Get Subscribing email list by email",
                "parameters": [
                    {
                        "description": "retrieve all email addresses that can receive update from an email address",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GetSubscribingEmailListRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/friends/subscribeFromEmail": {
            "post": {
                "description": "Requirement 4: As a user, I need an API to subscribe to updates from an email address.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend API"
                ],
                "summary": "Create a subscribe from email",
                "parameters": [
                    {
                        "description": "Subscribe to updates from an email address",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SubscribeRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/createUser": {
            "post": {
                "description": "Extend request: create a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "Create an User",
                "parameters": [
                    {
                        "description": "Create an User",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreatingUserRequest"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.BlockSubscribeRequest": {
            "type": "object",
            "properties": {
                "requestor": {
                    "type": "string"
                },
                "target": {
                    "type": "string"
                }
            }
        },
        "models.CommonFriendListRequest": {
            "type": "object",
            "properties": {
                "friends": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.CreatingUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "models.FriendConnectionRequest": {
            "type": "object",
            "properties": {
                "friends": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.FriendListRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "models.GetSubscribingEmailListRequest": {
            "type": "object",
            "properties": {
                "sender": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.SubscribeRequest": {
            "type": "object",
            "properties": {
                "requestor": {
                    "type": "string"
                },
                "target": {
                    "type": "string"
                }
            }
        }
    }
}